Created Windows_Timetable_notifier.py
This is a python script to notify daily college/school timetable on windows 10.
The program creates a timetable.db file and saves the timetable in it.
At the start of the every class added in the database the user would get notification on their windows 10 laptop stating which class has started and at what time will it end.